+++ components/camel-spring-xml/src/test/java/org/apache/camel/spring/processor/SpringConvertBodyTest.java
@@ -17,7 +17,7 @@
 package org.apache.camel.spring.processor;
 
 import org.apache.camel.CamelContext;
-import org.apache.camel.processor.ConvertBodyTest;
+import org.apache.camel.processor.converter.ConvertBodyTest;
 
 import static org.apache.camel.spring.processor.SpringTestHelper.createSpringCamelContext;
