+++ components/camel-spring-xml/src/test/java/org/apache/camel/spring/processor/SpringConvertHeaderTest.java
@@ -17,7 +17,7 @@
 package org.apache.camel.spring.processor;
 
 import org.apache.camel.CamelContext;
-import org.apache.camel.processor.ConvertHeaderTest;
+import org.apache.camel.processor.converter.ConvertHeaderTest;
 
 import static org.apache.camel.spring.processor.SpringTestHelper.createSpringCamelContext;
