+++ core/camel-core/src/test/java/org/apache/camel/processor/converter/custom/MyBean.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.tests.typeconverterscan;
+package org.apache.camel.processor.converter.custom;
 
 import org.apache.camel.util.ObjectHelper;
