+++ core/camel-core/src/test/resources/META-INF/services/org/apache/camel/TypeConverter
@@ -17,3 +17,4 @@
 
 org.apache.camel.converter.myconverter
 org.apache.camel.language.simple.myconverter
+org.apache.camel.processor.converter.custom
