+++ tests/pom.xml
@@ -41,6 +41,5 @@
         <module>camel-itest-standalone</module>
         <module>camel-itest</module>
         <module>camel-partial-classpath-test</module>
-        <module>camel-typeconverterscan-test</module>
     </modules>
 </project>
